refactor: move path_string to it's own file
