+++ src/lambda/evaluation_context.rs
@@ -1,17 +1,12 @@
-use std::borrow::Cow;
 use std::time::Duration;
 
 use async_graphql::dynamic::ResolverContext;
 #[allow(unused_imports)]
 use async_graphql::InputType;
-use async_graphql_value::ConstValue;
 use derive_setters::Setters;
 use reqwest::header::HeaderMap;
-use serde_json::json;
 
 use crate::http::RequestContext;
-use crate::json::JsonLike;
-use crate::request_template::AnyPath;
 
 // TODO: rename to ResolverContext
 #[derive(Clone, Setters)]
@@ -31,35 +26,6 @@ lazy_static::lazy_static! {
 impl Default for EvaluationContext<'static> {
   fn default() -> Self {
     Self::new(&REQUEST_CTX)
-  }
-}
-
-impl AnyPath for EvaluationContext<'_> {
-  fn any_path(&self, path: &[String]) -> Option<Cow<'_, str>> {
-    let ctx = self;
-    let resolver_ctx = ctx.context?;
-    let value = resolver_ctx.parent_value.as_value()?;
-    let mut result = None;
-    if let Some((head, tail)) = path.split_first() {
-      result = match head.as_str() {
-        "value" => value.get_path(tail).cloned(),
-        "args" => ctx.args()?.get_path(tail).cloned(),
-        "headers" => ctx.get_header_as_value(&tail[0]),
-        "vars" => Some(ConstValue::String(
-          ctx.req_ctx.server.vars.clone()?.get(&tail[0]).cloned()?,
-        )),
-        _ => None,
-      }
-      .and_then(|v| match v {
-        async_graphql::Value::String(s) => Some(Cow::Owned(s.clone())),
-        async_graphql::Value::Number(n) => Some(Cow::Owned(n.to_string())),
-        async_graphql::Value::Boolean(b) => Some(Cow::Owned(b.to_string())),
-        async_graphql::Value::Object(map) => Some(json!(map).to_string().into()),
-        async_graphql::Value::List(list) => Some(json!(list).to_string().into()),
-        _ => None,
-      });
-    }
-    result
   }
 }
